feature: implement new datatable, some pages that haven't implemented this might get error
