+++ app/Http/Resources/EmployeeResource.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace App\Http\Resources;
+
+use App\Models\Employee;
+use App\Http\Resources\BranchResource;
+use Illuminate\Http\Resources\Json\JsonResource;
+
+class EmployeeResource extends JsonResource
+{
+    /**
+     * Transform the resource collection into an array.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return array|\Illuminate\Contracts\Support\Arrayable|\JsonSerializable
+     */
+    public function toArray($request)
+    {
+        return [
+            'id' => $this->id,
+            'branch_id' => $this->branch_id,
+            'position_id' => $this->position_id,
+            'employee_id' => $this->employee_id,
+            'name' => $this->name,
+            'email' => $this->email,
+            'gender' => $this->gender,
+            'birth_date' => $this->birth_date,
+            'hiring_date' => $this->hiring_date,
+            'branches' => $this->branches,
+            'positions' => $this->positions
+        ];
+    }
+}
