+++ tailwind.config.js
@@ -18,5 +18,5 @@ module.exports = {
     },
   },
 
-  plugins: [require("@tailwindcss/forms"), require("flowbite/plugin")],
+  plugins: [require("@tailwindcss/forms")],
 };
